etl/requirements.txt
pandas
yfinance
supabase
python-dotenv
